+++ packages/react-core/src/components/molecules/LabeledSwitch/LabeledSwitch.tsx
@@ -2,8 +2,16 @@ import { SwitchProps } from '../../atoms/Switch';
 
 export type LabelPositionOptions = 'left' | 'right';
 
-export interface LabeledSwitchProps extends SwitchProps {
-  label: string;
-  labelPosition?: LabelPositionOptions;
+export type LabeledSwitchProps = {
   pressableLabel?: boolean;
-}
+  leftLabel?: string;
+  rightLabel?: string;
+} & (
+  | {
+      leftLabel: string;
+    }
+  | {
+      rightLabel: string;
+    }
+) &
+  SwitchProps;